app/pyproject.toml
[tool.poetry]
name = "spotify-bpm-checker"
version = "0.1.0"
description = ""
authors = ["Marcus Dragsten <marcuskd@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.4.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
